backend/voice_agent/requirements.txt
# Voice agent dependencies (Fly.io image).
#
# Superset of backend/requirements.txt + voice-specific SDKs. Keeping
# one requirements file per image (rather than -r ../requirements.txt)
# because the Dockerfile build context is this directory only.

# --- Existing backend deps (mirror backend/requirements.txt) ---
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
anthropic
stripe
python-dotenv
google-auth
google-generativeai
requests
eval-type-backport
bcrypt
python-jose[cryptography]
python-multipart
twilio
reportlab

# --- Voice-specific ---
livekit-agents
livekit-api
livekit-plugins-deepgram
livekit-plugins-cartesia
livekit-plugins-elevenlabs
deepgram-sdk
cartesia
elevenlabs

# Direct HTTP+SSE client for the custom Eleven v3 adapter.
# livekit-plugins-elevenlabs forces WebSockets, which v3 doesn't support
# (returns 403). See backend/voice_agent/elevenlabs_v3_tts.py for context.
aiohttp
